+++ charting/src/main/java/com/zygne/chart/chart/model/chart/Colors.java
@@ -5,9 +5,14 @@ public final class Colors {
     public static final String RED = "#C34F3F";
     public static final String GREEN = "#45B357";
     public static final String BLUE = "#00B2EB";
-    public static final String WHITE = "#EFF5F0";
+    public static final String BLUE_DARK = "#02123B";
+    public static final String SNOW_WHITE = "#EFF5F0";
     public static final String YELLOW = "#C3C33F";
     public static final String ORANGE = "#D9B333";
+    public static final String ORANGE_BRIGHT = "#FF5733";
+    public static final String WHITE = "#FFFFFF";
+    public static final String BLACK = "#000000";
+    public static final String SILVER = "#B5BCCE";
 
     private Colors(){}
 }
