+++ charting/src/main/java/com/zygne/chart/chart/model/chart/Object2d.java
@@ -30,6 +30,7 @@ public abstract class Object2d extends BoxCollider implements Renderable {
         BLUE,
         GREEN,
         WHITE,
-        HIGHLIGHT
+        HIGHLIGHT,
+        GRAY
     }
 }
