+++ charting/src/main/java/com/zygne/chart/chart/model/data/BarSerie.java
@@ -3,6 +3,7 @@ package com.zygne.chart.chart.model.data;
 public class BarSerie extends Serie {
 
     private double value;
+    private boolean included;
 
     public BarSerie(Double value){
         this.value = value;
@@ -14,5 +15,13 @@ public class BarSerie extends Serie {
 
     public void setValue(double value) {
         this.value = value;
+    }
+
+    public boolean isIncluded() {
+        return included;
+    }
+
+    public void setIncluded(boolean included) {
+        this.included = included;
     }
 }
