--- data/src/main/java/com/zygne/data/domain/model/VolumePrice.java
@@ -1,12 +1,0 @@
-package com.zygne.data.domain.model;
-
-public record VolumePrice(double price, long size) {
-
-    public double getPrice() {
-        return price;
-    }
-
-    public long getSize() {
-        return size;
-    }
-}
