--- data/src/main/java/com/zygne/data/domain/model/VolumePriceSum.java
@@ -1,11 +1,0 @@
-package com.zygne.data.domain.model;
-
-public class VolumePriceSum {
-    public final double price;
-    public long totalSize;
-
-    public VolumePriceSum(double price, long totalSize) {
-        this.price = price;
-        this.totalSize = totalSize;
-    }
-}
