+++ data/src/main/java/com/zygne/data/presentation/presenter/implementation/MainPresenterImpl.java
@@ -2,21 +2,17 @@ package com.zygne.data.presentation.presenter.implementation;
 
 import com.zygne.data.YahooDataBroker;
 import com.zygne.data.domain.DataBroker;
-import com.zygne.data.domain.interactor.implementation.data.base.VolumePriceInteractor;
 import com.zygne.data.domain.model.Histogram;
-import com.zygne.data.domain.model.LiquidityLevel;
 import com.zygne.data.presentation.presenter.base.MainPresenter;
 import com.zygne.arch.domain.Logger;
 import com.zygne.arch.domain.executor.Executor;
 import com.zygne.arch.domain.executor.MainThread;
 import com.zygne.arch.presentation.presenter.base.BasePresenter;
 import com.zygne.data.presentation.presenter.implementation.flow.DataFlow;
-import com.zygne.data.presentation.presenter.implementation.flow.SupplyFlow;
 
 import java.util.List;
 
 public class MainPresenterImpl extends BasePresenter implements MainPresenter,
-        SupplyFlow.Callback,
         DataFlow.Callback {
 
     private final DataBroker dataBroker;
@@ -25,7 +21,6 @@ public class MainPresenterImpl extends BasePresenter implements MainPresenter,
     private List<Histogram> histogramList;
     private boolean downloadingData = false;
     private final DataFlow dataFlow;
-    private final SupplyFlow supplyFlow;
     private String dateRange = "";
 
     public MainPresenterImpl(Executor executor, MainThread mainThread, MainPresenter.View view, Logger logger) {
@@ -34,7 +29,6 @@ public class MainPresenterImpl extends BasePresenter implements MainPresenter,
         this.dataBroker = new YahooDataBroker(logger);
         this.view = view;
         this.view.showError("");
-        this.supplyFlow = new SupplyFlow(executor, mainThread, this);
         this.dataFlow = new DataFlow(executor, mainThread, this, logger);
 
         view.prepareView();
@@ -63,17 +57,12 @@ public class MainPresenterImpl extends BasePresenter implements MainPresenter,
     }
 
     @Override
-    public void onSupplyCompleted(List<LiquidityLevel> liquidityLevels) {
-        downloadingData = false;
-        view.hideLoading();
-        view.onComplete(histogramList, liquidityLevels, ticker, dateRange);
-    }
-
-    @Override
     public void onDataFetched(List<Histogram> data, String time) {
+        this.downloadingData = false;
         this.dateRange = time;
         this.histogramList = data;
-        supplyFlow.start(data, VolumePriceInteractor.PriceStructure.OHLCM);
+        view.hideLoading();
+        view.onComplete(histogramList, ticker, dateRange);
     }
 
     @Override
